Remove the r2 example
travis' Ubuntu xenial does not have an up to date enough r2 so the
script fails
