+++ .travis.yml
@@ -8,15 +8,9 @@ os:
 python:
 - 3.6
 
-addons:
-  apt:
-    update: true
-
 install:
   - pip install manticore[native]
   - pip install nose
-  - pip install r2pipe
-  - sudo apt install radare2
 
 script:
   - nosetests
